# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Nova Mart Commerce API
  version: 1.0.0
  description: |
    Public HTTP contract for the **Nova Mart** online shopping platform.

    This document is the **source of truth**. It was authored *before* any service
    implementation, and every backend endpoint and every frontend API client method
    is derived from it. Changing behaviour means changing this file first.

    ## Architecture

    Clients never talk to a service directly. All traffic enters through the
    **API Gateway** (`/api/v1/**`), which authenticates the caller, strips any
    client-supplied identity headers, injects trusted ones, and routes to the
    owning service.

    | Path prefix                                   | Owning service       | Owns database     |
    | --------------------------------------------- | -------------------- | ----------------- |
    | `/auth`, `/users`                              | auth-service         | `auth_db`         |
    | `/products`, `/categories`, `/brands`          | product-service      | `product_db`      |
    | `/cart`                                        | cart-service         | `cart_db`         |
    | `/orders`                                      | order-service        | `order_db`        |
    | `/payments`                                    | payment-service      | `payment_db`      |
    | `/inventory`                                   | inventory-service    | `inventory_db`    |
...

...

    ## Response envelope

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /cart:
    get:
      tags: [Cart]
      operationId: getCart
      summary: Get the caller's cart
      description: Returns an empty cart rather than 404 when the shopper has never added anything.
      responses:
        '200':
          description: Cart with computed totals
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Cart'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Cart]
      operationId: clearCart
      summary: Remove every line from the cart
      responses:
        '204':
          description: Cart emptied
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cart/items:
    post:
      tags: [Cart]
      operationId: addCartItem
      summary: Add a product to the cart
      description: |
        Adding a product already present increments the existing line rather than
        creating a duplicate. Stock availability is checked against inventory-service
        before the line is written.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
            example:
              productId: 8f14e45f-ea8f-4b2a-9d1c-3c2b7a51de90
              quantity: 1
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Cart'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No such product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Requested quantity exceeds available stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: Only 3 units of this item are available
                errorCode: INSUFFICIENT_STOCK
                timestamp: '2026-02-14T10:10:00.000Z'
                path: /api/v1/cart/items
  /cart/items/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Cart]
      operationId: updateCartItem
      summary: Set the quantity of a cart line
      description: Sending `quantity` of 0 removes the line.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Cart'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Requested quantity exceeds available stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Cart]
      operationId: removeCartItem
      summary: Remove a line from the cart
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Cart'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # =====================================================================
  #  ORDERS
  # =====================================================================
...
  /inventory:
    get:
      tags: [Inventory]
      operationId: listInventory
      summary: List stock records (admin)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: lowStockOnly
          in: query
          description: Return only records where available quantity is at or below the reorder threshold
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Paged stock records
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedInventory'
        '403':
          $ref: '#/components/responses/Forbidden'
  /inventory/batch:
    post:
      tags: [Inventory]
      operationId: getInventoryBatch
      summary: Check availability for many products at once
      description: |
        Added to the contract when the cart page needed live stock for every line.
        Asking product by product would issue one HTTP round trip per line, which
        is the N+1 problem moved onto the network where each hop costs
        milliseconds rather than microseconds. One call bounds it at a constant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productIds]
              properties:
                productIds:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Stock records for the ids that exist. Unknown ids are omitted rather than returned as errors.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/InventoryItem'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /inventory/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Inventory]
      operationId: getInventory
      summary: Check availability for one product
      security: []
      responses:
        '200':
          description: Stock record
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InventoryItem'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Inventory]
      operationId: updateInventory
      summary: Set the on-hand quantity (admin)
      description: Records a `MANUAL_ADJUSTMENT` stock transaction for the audit trail.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInventoryRequest'
      responses:
        '200':
          description: Stock updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InventoryItem'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /inventory/{productId}/reserve:
    parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Inventory]
      operationId: reserveStock
      summary: Hold stock for an in-flight order
      description: |
        Moves units from `availableQuantity` to `reservedQuantity` under a
        row-level lock, so two shoppers cannot reserve the same last unit.
        Called by order-service during checkout.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockMovementRequest'
      responses:
        '200':
          description: Stock reserved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InventoryItem'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Not enough available stock to satisfy the reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: Only 2 units available, 5 requested
                errorCode: INSUFFICIENT_STOCK
                timestamp: '2026-02-14T10:21:02.311Z'
                path: /api/v1/inventory/8f14.../reserve
  /inventory/{productId}/release:
    parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Inventory]
      operationId: releaseStock
      summary: Return reserved stock to the available pool
      description: The compensating action for `reserve`. Used when payment fails or an order is cancelled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockMovementRequest'
      responses:
        '200':
          description: Stock released
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InventoryItem'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Attempted to release more than is currently reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inventory/{productId}/commit:
    parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Inventory]
      operationId: commitStock
      summary: Consume a reservation once payment has settled
      description: |
        Permanently removes units from both `reservedQuantity` and `totalQuantity`.
        This is the third leg of the reserve / release / commit protocol and is what
        makes the checkout saga safe: stock leaves the warehouse only after money
        has actually been captured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockMovementRequest'
      responses:
        '200':
          description: Reservation consumed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/InventoryItem'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Attempted to commit more than is currently reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =====================================================================
  #  NOTIFICATIONS
  # =====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT issued by auth-service. Send as `Authorization: Bearer <accessToken>`.'
  parameters:
    Page:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 12
  responses:
    ValidationError:
      description: The request body or query string failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Request validation failed
            errorCode: VALIDATION_FAILED
            timestamp: '2026-02-14T09:40:12.882Z'
            path: /api/v1/auth/register
            fieldErrors:
              - field: password
                message: Password must be at least 8 characters
              - field: email
                message: Must be a well-formed email address
    Unauthorized:
      description: Missing, malformed or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Authentication is required to access this resource
            errorCode: UNAUTHORIZED
            timestamp: '2026-02-14T09:41:03.114Z'
            path: /api/v1/cart
    Forbidden:
      description: Authenticated, but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: You do not have permission to perform this action
            errorCode: FORBIDDEN
            timestamp: '2026-02-14T09:42:55.900Z'
            path: /api/v1/products
...
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Product not found
            errorCode: PRODUCT_NOT_FOUND
            timestamp: '2026-02-14T09:43:30.720Z'
            path: /api/v1/products/8f14e45f-ea8f-4b2a-9d1c-3c2b7a51de90
  schemas:
...
    ApiResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Product retrieved successfully
        data:
          description: Operation-specific payload. Absent for `204` responses.
    ErrorResponse:
      type: object
      required: [success, message, errorCode, timestamp]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
          description: Human-readable, safe to show a shopper. Never contains a stack trace.
        errorCode:
          type: string
          description: Stable machine-readable identifier. Clients branch on this, not on the message text.
          example: PRODUCT_NOT_FOUND
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        fieldErrors:
          type: array
          description: Present only when `errorCode` is `VALIDATION_FAILED`.
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages, first, last]
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 12
        totalElements:
          type: integer
          format: int64
          example: 24
        totalPages:
          type: integer
          example: 2
        first:
          type: boolean
        last:
          type: boolean

    # ---------- auth ----------
...
    Cart:
      type: object
      required: [id, userId, items, subtotal, totalQuantity, currency]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: number
          format: double
          description: Sum of line totals. Delivery and discounts are applied at checkout, not here.
        totalQuantity:
          type: integer
        currency:
          type: string
          example: INR
        updatedAt:
          type: string
          format: date-time
    CartItem:
      type: object
      required: [productId, name, slug, imageUrl, unitPrice, quantity, lineTotal]
      properties:
        productId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
          format: uri
        unitPrice:
          type: number
          format: double
          description: Refreshed from product-service on every read, so a price change is visible before checkout.
        quantity:
          type: integer
          minimum: 1
        lineTotal:
          type: number
          format: double
        availableQuantity:
          type: integer
          description: Live stock, so the cart page can warn before the shopper reaches checkout.
        inStock:
          type: boolean
    AddCartItemRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 20
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 0
          maximum: 20
          description: Zero removes the line.

    # ---------- orders ----------
...
    InventoryItem:
      type: object
      required: [productId, totalQuantity, reservedQuantity, availableQuantity, inStock]
      properties:
        productId:
          type: string
          format: uuid
        totalQuantity:
          type: integer
          description: Physical units held.
        reservedQuantity:
          type: integer
          description: Units held for orders that are in flight but not yet settled.
        availableQuantity:
          type: integer
          description: totalQuantity minus reservedQuantity. This is what a shopper may buy.
        reorderThreshold:
          type: integer
          description: Availability at or below this level marks the product low stock in the admin dashboard.
          example: 5
        inStock:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    StockMovementRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
        referenceId:
          type: string
          description: Order id, recorded on the stock transaction for traceability.
          nullable: true
    UpdateInventoryRequest:
      type: object
      required: [totalQuantity]
      properties:
        totalQuantity:
          type: integer
          minimum: 0
        reorderThreshold:
          type: integer
          minimum: 0
    PagedInventory:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InventoryItem'
        page:
          $ref: '#/components/schemas/PageMeta'

    # ---------- notifications ----------
...